backend/services/analysis/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
minio==7.2.0
python-multipart==0.0.6
python-magic==0.4.27